chore: update contract migrations after 3.14 release
