+++ crates/contract-history/src/lib.rs
@@ -1,9 +1,9 @@
 pub const fn current_mainnet() -> &'static [u8] {
-    version_3_13_0()
+    version_3_14_0()
 }
 
 pub const fn current_testnet() -> &'static [u8] {
-    version_3_13_0()
+    version_3_14_0()
 }
 
 pub const fn version_2_2_0() -> &'static [u8; 566653] {
@@ -68,6 +68,10 @@ pub const fn version_3_12_0() -> &'static [u8; 1482860] {
 
 pub const fn version_3_13_0() -> &'static [u8; 1509901] {
     include_bytes!("../archive/signer-3_13_0.wasm")
+}
+
+pub const fn version_3_14_0() -> &'static [u8; 1187525] {
+    include_bytes!("../archive/signer-3_14_0.wasm")
 }
 
 #[cfg(test)]
