+++ crates/contract/src/v3_14_0_state.rs
@@ -1,0 +1,75 @@
+//! ## Overview
+//! This module stores the previous contract state—the one you want to migrate from.
+//! The goal is to describe the data layout _exactly_ as it existed before.
+//!
+//! ## Guideline
+//! In theory, you could copy-paste every struct from the specific commit you're migrating from.
+//! However, this approach (a) requires manual effort from a developer and (b) increases the binary size.
+//! A better approach: only copy the structures that have changed and import the rest from the existing codebase.
+
+use borsh::{BorshDeserialize, BorshSerialize};
+use near_mpc_contract_interface::types::{Metrics, VerifyForeignTransactionRequest};
+use near_sdk::{
+    AccountId, env,
+    store::{Lazy, LookupMap},
+};
+
+use crate::{
+    SupportedForeignChainsByNode,
+    config::Config,
+    foreign_chains_metadata::ForeignChainsMetadata,
+    node_migrations::NodeMigrations,
+    primitives::{
+        ckd::CKDRequest,
+        signature::{SignatureRequest, YieldIndex},
+    },
+    state::ProtocolContractState,
+    tee::{tee_state::TeeState, verifier_votes::TeeVerifierVotes},
+    update::ProposedUpdates,
+};
+
+/// Keep this module in sync with [`crate::MpcContract`]: the moment a field's borsh
+/// layout diverges, shadow the old type here (see this module's history for examples) so
+/// state written by the `3.14.0` contract still deserializes during migration.
+#[derive(Debug, BorshSerialize, BorshDeserialize)]
+pub struct MpcContract {
+    protocol_state: ProtocolContractState,
+    pending_signature_requests: LookupMap<SignatureRequest, Vec<YieldIndex>>,
+    pending_ckd_requests: LookupMap<CKDRequest, Vec<YieldIndex>>,
+    pending_verify_foreign_tx_requests: LookupMap<VerifyForeignTransactionRequest, Vec<YieldIndex>>,
+    proposed_updates: ProposedUpdates,
+    node_foreign_chain_support: SupportedForeignChainsByNode,
+    config: Config,
+    tee_state: TeeState,
+    accept_requests: bool,
+    node_migrations: NodeMigrations,
+    metrics: Metrics,
+    foreign_chains: Lazy<ForeignChainsMetadata>,
+    tee_verifier_account_id: Option<AccountId>,
+    tee_verifier_votes: TeeVerifierVotes,
+}
+
+impl From<MpcContract> for crate::MpcContract {
+    fn from(old: MpcContract) -> Self {
+        if !matches!(old.protocol_state, ProtocolContractState::Running(_)) {
+            env::panic_str("Contract must be in running state when migrating.");
+        }
+
+        crate::MpcContract {
+            protocol_state: old.protocol_state,
+            pending_signature_requests: old.pending_signature_requests,
+            pending_ckd_requests: old.pending_ckd_requests,
+            pending_verify_foreign_tx_requests: old.pending_verify_foreign_tx_requests,
+            proposed_updates: old.proposed_updates,
+            node_foreign_chain_support: old.node_foreign_chain_support,
+            config: old.config,
+            tee_state: old.tee_state,
+            accept_requests: old.accept_requests,
+            node_migrations: old.node_migrations,
+            metrics: old.metrics,
+            foreign_chains: old.foreign_chains,
+            tee_verifier_account_id: old.tee_verifier_account_id,
+            tee_verifier_votes: old.tee_verifier_votes,
+        }
+    }
+}
